Create CV
A Curriculum Vitae, or CV for short, is a professional document that summarizes your work history, education, and skills. The main purpose of a CV is to sell you, as a candidate, to prospective employers. When applying for a position, you're always going to be asked for a CV, coupled with a cover letter.